Better abstraction for storage subsystem
1. In addition to commit 91e5211, renamed swSubsystem::filesystem namespace to
internalHardwareSubsystem::storage because the internal flash can support several
filesystem types including Key-value stores(NVS type filesystem in espressif lingo)
as well as the SPI Flash filesystem(SPIFFS) among others such as LittleFS.
2. (1) above allows greater visibility at an overall system level that we are using
the internal flash as opposed to an external store such as an SD card. This can give
further visibility that we are using utilities specific to a given SOC such as the
ESP32 which may need to be modified to suit a different SoC when porting.