Change relation between PM sensor and loadswitch GPIO to composition
1. Changed relation of the above from an generalization relation -
"PM sensor implements power switch interface",thereby having the PM sensor
as a subclass to the GPIO class, to a composition relationship -
"PM sensor has an instance of power switch interface and the power switch
interface cannot exist - from a mental model -  without a load such as a
particulate sensor"
2. The loadwitch is onlu given relevance based on the loads. This change means
that the loadswitch cannot exist without the loads.
3. This would allow the loadswitch to be called by referring to a given load
such as particulateSensor.powerSwitch.on() as it would not make sense from
a mental model for a loadswitch to exist in isolation without loads.
4. At a PCB level, the loadswitch only has the particulate sensor downstream
therefore the multiplicity on a loadswitch to load relation is 1:1 on the
current board revision. However, this may change in future with a need of
such a change being transparent from a software model perspective.