Prevent possible image buffer modification
Added compile time checks to prevent image buffer modification
by a non const pointer.

This enforces a compile time contract that, any receiver of a
pointer to the image buffer through public functions getImage() or
getAndPrintImage() will enforce that the pointer will not change the
contents of the image buffer by enforcing a mandatory const modifier
on the lvalue.