Improve chamber test plotting method
Minimized disk writes by allowing the plots to be made from RAM at
the expense of a greater RAM footprint. This is due to the fact that
the notebook may be run on a raspberry pi which greatly suffers MicroSD
failures. The data is persisted to Disk as a CSV file whenever the ipython
kernel is interrupted.

The primary tool this will be run on is a raspberry pi 4B which has
8GB of available RAM. This compromise is therefore suitable for both
desktop and single board computer systems.