Reduce buffer size in UART driver
1. This is because of the nature the UART receive timeout is
implemented. The driver seems not to respond to the timeout.
In terms of expected behaviour, data reception should terminate
when either the user buffer is full OR the timeout value has been
reached. Observed behaviour is that the timeout value is ignored and
the device keeps collecting data until the buffer is full. As such
the time needed to perform data acquisition gets longer especially
when working in a sequential manner.
2. I hypothesize the timeout value is passed to the lower driver layers
to represent timeout on UART bus inactivity. Given the particulate
sensor in active reporting mode keeps sending data, the timeout is ignored
until the user buffer is full.
3. A solution for this to reduce processing time is to handle UART data reception
in a non-sequential manner by placing the work of acquisition on a task. Once the task
completes, it sends via an IPC data that signals that acquisition is complete.
This would mean that the functions would need to be split at the driver level.